+++ repository/types.go
@@ -6,7 +6,7 @@ type UserRepository interface {
 	Create(user *model.User) (*model.User, error)
 	Delete(id int64) error
 	GetPasswordHash(id int64) (string, error)
-	GetStatus(id int64) (uint8, error)
+	GetStatus(id int64) (int, error)
 	GetByID(id int64) (*model.User, error)
 	GetByUsername(username string) (*model.User, error)
 	UpdatePasswordHash(id int64, newHash string) error
